Create readCSV_and_generateReport.py
First, this script reads a CSV file containing a list of employees and their departments. 
Second, it turns the CSV data into a list of dictionaries, stored in a variable in this script.
Third, this script iterates through each item in the list to find the names of each department and counts them.
Lastly, it creates a new text file showing each department and the number of people in them (by the count of how many times the department name came up).